CHANGE IN API, better precision
* Now, height and width is specified in 'double'.
* Precision is increased to 4 digits

+++ inc/emf2svg.h
@@ -19,9 +19,9 @@ typedef struct {
     // draw svg document delimiter or not
     bool svgDelimiter;
     // height of the target image
-    int imgHeight;
+    double imgHeight;
     // width of the target image
-    int imgWidth;
+    double imgWidth;
 } generatorOptions;
 
 // covert function
